Move ToolBar inline
