+++ src/sources.cmake
@@ -165,7 +165,6 @@ target_sources(${PROJECT_NAME} PRIVATE
     libtcod/gui/statusbar.hpp
     libtcod/gui/textbox.hpp
     libtcod/gui/togglebutton.hpp
-    libtcod/gui/toolbar.cpp
     libtcod/gui/toolbar.hpp
     libtcod/gui/vbox.hpp
     libtcod/gui/widget.hpp
@@ -460,7 +459,6 @@ source_group(libtcod\\gui FILES
     libtcod/gui/statusbar.hpp
     libtcod/gui/textbox.hpp
     libtcod/gui/togglebutton.hpp
-    libtcod/gui/toolbar.cpp
     libtcod/gui/toolbar.hpp
     libtcod/gui/vbox.hpp
     libtcod/gui/widget.hpp
